feat: adiciona permissões de usuário referentes ao serviço de escolas
